feat: add hcl parsing of stack.id
