+++ hcl/hcl.go
@@ -17,6 +17,7 @@ package hcl
 import (
 	"os"
 	"path/filepath"
+	"regexp"
 	"sort"
 	"strings"
 
@@ -78,8 +79,16 @@ type Terramate struct {
 	Config *RootConfig
 }
 
+// StackID represents the stack ID. Its zero value represents an undefined ID.
+type StackID struct {
+	id *string
+}
+
 // Stack is the parsed "stack" HCL block.
 type Stack struct {
+	// ID of the stack. If the ID is nil it indicates this stack has no ID.
+	ID StackID
+
 	// Name of the stack
 	Name string
 
@@ -144,6 +153,17 @@ type TerramateParser struct {
 
 	// Blocks are the unmerged blocks from all files.
 	Blocks ast.Blocks
+}
+
+var stackIDRegex = regexp.MustCompile("^[a-zA-Z0-9_-]{1,64}$")
+
+// NewStackID creates a new StackID with the given string as its id.
+// It guarantees that the id passed is a valid StackID, an error is returned otherwise.
+func NewStackID(id string) (StackID, error) {
+	if !stackIDRegex.MatchString(id) {
+		return StackID{}, errors.E("Stack ID %q doesn't match %q", id, stackIDRegex)
+	}
+	return StackID{id: &id}, nil
 }
 
 // NewTerramateParser creates a Terramate parser for the directory dir.
@@ -674,6 +694,20 @@ func parseStack(stack *Stack, stackblock *ast.Block) error {
 			Msg("Setting attribute on configuration.")
 
 		switch attr.Name {
+		case "id":
+			if attrVal.Type() != cty.String {
+				errs.Append(errors.E(attr.NameRange,
+					"field stack.\"id\" must be a \"string\" but is %q",
+					attrVal.Type().FriendlyName()),
+				)
+				continue
+			}
+			id, err := NewStackID(attrVal.AsString())
+			if err != nil {
+				errs.Append(errors.E(attr.NameRange, err))
+				continue
+			}
+			stack.ID = id
 		case "name":
 			if attrVal.Type() != cty.String {
 				errs.Append(errors.E(attr.NameRange,
