+++ hcl/hcl_test.go
@@ -419,459 +419,6 @@ func TestHCLParserRootConfig(t *testing.T) {
 	}
 }
 
-func TestHCLParserStack(t *testing.T) {
-	for _, tc := range []testcase{
-		{
-			name: "empty stack block",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-						stack {}
-					`,
-				},
-			},
-			want: want{
-				config: hcl.Config{
-					Terramate: &hcl.Terramate{},
-					Stack:     &hcl.Stack{},
-				},
-			},
-		},
-		{
-			name: "stack with unrecognized blocks",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						stack{
-							block1 {}
-							block2 {}
-						}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrTerramateSchema),
-					errors.E(hcl.ErrTerramateSchema),
-				},
-			},
-		},
-		{
-			name: "multiple stack blocks",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {}
-						stack{}
-						stack{}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrTerramateSchema),
-				},
-			},
-		},
-		{
-			name: "empty name",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-						stack {
-							name = ""
-						}
-					`,
-				},
-			},
-			want: want{
-				config: hcl.Config{
-					Terramate: &hcl.Terramate{},
-					Stack:     &hcl.Stack{},
-				},
-			},
-		},
-		{
-			name: "name is not a string - fails",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-						stack {
-							name = 1
-						}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrTerramateSchema,
-						mkrange("stack.tm", start(6, 8, 77), end(6, 12, 81)),
-					),
-				},
-			},
-		},
-		{
-			name: "name has interpolation - fails",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-						stack {
-							name = "${test}"
-						}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrTerramateSchema,
-						mkrange("stack.tm", start(6, 18, 87), end(6, 22, 91))),
-					errors.E(hcl.ErrTerramateSchema,
-						mkrange("stack.tm", start(6, 8, 77), end(6, 12, 81))),
-				},
-			},
-		},
-		{
-			name: "unrecognized attribute name - fails",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-						stack {
-							bleh = "a"
-						}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrTerramateSchema),
-				},
-			},
-		},
-		{
-			name: "schema not checked for files with syntax errors",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-						stack {
-							wants =
-							unrecognized = "test"
-						}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrHCLSyntax),
-				},
-			},
-		},
-		{
-			name: "after: empty set works",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-						stack {}
-					`,
-				},
-			},
-			want: want{
-				config: hcl.Config{
-					Terramate: &hcl.Terramate{},
-					Stack:     &hcl.Stack{},
-				},
-			},
-		},
-		{
-			name: "'after' single entry",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-
-						stack {
-							after = ["test"]
-						}
-					`,
-				},
-			},
-			want: want{
-				config: hcl.Config{
-					Terramate: &hcl.Terramate{},
-					Stack: &hcl.Stack{
-						After: []string{"test"},
-					},
-				},
-			},
-		},
-		{
-			name: "'after' invalid element entry",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						stack {
-							after = [1]
-						}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrTerramateSchema),
-				},
-			},
-		},
-		{
-			name: "'after' invalid type",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						stack {
-							after = {}
-						}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrTerramateSchema),
-				},
-			},
-		},
-		{
-			name: "'after' null value",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						stack {
-							after = null
-						}
-					`,
-				},
-			},
-			want: want{
-				config: hcl.Config{
-					Stack: &hcl.Stack{
-						After: []string{},
-					},
-				},
-			},
-		},
-		{
-			name: "'after' duplicated entry",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						stack {
-							after = ["test", "test"]
-						}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrTerramateSchema),
-				},
-			},
-		},
-		{
-			name: "multiple 'after' fields - fails",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						stack {
-							after = ["test"]
-							after = []
-						}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrHCLSyntax),
-				},
-			},
-		},
-		{
-			name: "multiple 'before' fields - fails",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						stack {
-							before = []
-							before = []
-						}
-					`,
-				},
-			},
-			want: want{
-				errs: []error{
-					errors.E(hcl.ErrHCLSyntax),
-				},
-			},
-		},
-		{
-			name: "'before' single entry",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-
-						stack {
-							before = ["something"]
-						}
-					`,
-				},
-			},
-			want: want{
-				config: hcl.Config{
-					Terramate: &hcl.Terramate{},
-					Stack: &hcl.Stack{
-						Before: []string{"something"},
-					},
-				},
-			},
-		},
-		{
-			name: "'before' multiple entries",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-
-						stack {
-							before = ["something", "something-else", "test"]
-						}
-					`,
-				},
-			},
-			want: want{
-				config: hcl.Config{
-					Terramate: &hcl.Terramate{},
-					Stack: &hcl.Stack{
-						Before: []string{"something", "something-else", "test"},
-					},
-				},
-			},
-		},
-		{
-			name: "stack with valid description",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						stack {
-							description = "some cool description"
-						}
-					`,
-				},
-			},
-			want: want{
-				config: hcl.Config{
-					Stack: &hcl.Stack{
-						Description: "some cool description",
-					},
-				},
-			},
-		},
-		{
-			name: "stack with multiline description",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-					stack {
-						description =  <<-EOD
-	line1
-	line2
-	EOD
-					}`,
-				},
-			},
-			want: want{
-				config: hcl.Config{
-					Stack: &hcl.Stack{
-						Description: "line1\nline2",
-					},
-				},
-			},
-		},
-		{
-			name: "'before' and 'after'",
-			input: []cfgfile{
-				{
-					filename: "stack.tm",
-					body: `
-						terramate {
-							required_version = ""
-						}
-
-						stack {
-							before = ["something"]
-							after = ["else"]
-						}
-					`,
-				},
-			},
-			want: want{
-				config: hcl.Config{
-					Terramate: &hcl.Terramate{},
-					Stack: &hcl.Stack{
-						Before: []string{"something"},
-						After:  []string{"else"},
-					},
-				},
-			},
-		},
-	} {
-		testParser(t, tc)
-	}
-}
-
 func TestHCLParserMultipleErrors(t *testing.T) {
 	for _, tc := range []testcase{
 		{
